+++ app/src/main/java/cn/spacexc/wearbili/dataclass/bangumi/Pendant.kt
@@ -3,5 +3,5 @@ package cn.spacexc.wearbili.dataclass.bangumi
 data class Pendant(
     val image: String,
     val name: String,
-    val pid: Int
+    val pid: Long
 )
