+++ app/src/main/java/cn/spacexc/wearbili/dataclass/bangumi/Section.kt
@@ -2,10 +2,10 @@ package cn.spacexc.wearbili.dataclass.bangumi
 
 data class Section(
     val attr: Int,
-    val episode_id: Int,
+    val episode_id: Long,
     val episode_ids: List<Any>,
     val episodes: List<EpisodeX>,
-    val id: Int,
+    val id: Long,
     val title: String,
     val type: Int
 )
